:recycle: refactor!: add abstract manager
fixes error when the player tries to teleport to a pocket or linker after server restart
