+++ src/main/java/dev/ipoleksenko/PocketHome/PocketHomePlugin.java
@@ -1,5 +1,6 @@
 package dev.ipoleksenko.PocketHome;
 
+import dev.ipoleksenko.PocketHome.manager.LinkerManager;
 import dev.ipoleksenko.PocketHome.manager.PocketManager;
 import org.bukkit.plugin.java.JavaPlugin;
 import org.jetbrains.annotations.Contract;
@@ -11,6 +12,7 @@ public class PocketHomePlugin extends JavaPlugin {
 	private static final Integer POCKET_RADIUS = 2; // in chunks; generate (2 * n)^2 chunks
 	private static PocketHomePlugin instance;
 	private PocketManager pocketManager;
+	private LinkerManager linkerManager;
 
 	/**
 	 * Get PocketHome plugin instance
@@ -39,15 +41,27 @@ public class PocketHomePlugin extends JavaPlugin {
 	 * Get PocketManager instance
 	 *
 	 * @return PocketManager instance
+	 * @see PocketManager
 	 */
 	public PocketManager getPocketManager() {
 		return this.pocketManager;
 	}
 
+	/**
+	 * Get LinkerManager instance
+	 *
+	 * @return LinkerManager instance
+	 * @see LinkerManager
+	 */
+	public LinkerManager getLinkerManager() {
+		return this.linkerManager;
+	}
+
 	@Override
 	public void onEnable() {
 		instance = this;
-		pocketManager = new PocketManager();
+		this.pocketManager = new PocketManager();
+		this.linkerManager = new LinkerManager();
 	}
 
 	@Override
